+++ CHANGELOG.md
@@ -253,3 +253,4 @@
 - 2024-01-15: Add qualify, main, deal, campaigns, +4 types
 - 2024-01-16: Features sender, follow up, campaign, main, +6
 - 2024-01-17: Init campaign, chat, mutate, descend, +6
+- 2024-01-18: Refactor README, follow up, campaigns, bettercontact, +6